// src/app/constants/item-type.enum.ts
export enum ItemTypeEnum {
  weapon = 'weapon',
  weaponGrade = 'weaponGrade',
  weaponRefine = 'weaponRefine',
  weaponCard1 = 'weaponCard1',
  weaponCard2 = 'weaponCard2',
  weaponCard3 = 'weaponCard3',
  weaponCard4 = 'weaponCard4',
  weaponEnchant0 = 'weaponEnchant0',
  weaponEnchant1 = 'weaponEnchant1',
  weaponEnchant2 = 'weaponEnchant2',
  weaponEnchant3 = 'weaponEnchant3',

  ammo = 'ammo',

  leftWeapon = 'leftWeapon',
  leftWeaponRefine = 'leftWeaponRefine',
  leftWeaponGrade = 'leftWeaponGrade',
  leftWeaponCard1 = 'leftWeaponCard1',
  leftWeaponCard2 = 'leftWeaponCard2',
  leftWeaponCard3 = 'leftWeaponCard3',
  leftWeaponCard4 = 'leftWeaponCard4',
  leftWeaponEnchant0 = 'leftWeaponEnchant0',
  leftWeaponEnchant1 = 'leftWeaponEnchant1',
  leftWeaponEnchant2 = 'leftWeaponEnchant2',
  leftWeaponEnchant3 = 'leftWeaponEnchant3',

  shield = 'shield',
  shieldRefine = 'shieldRefine',
  shieldGrade = 'shieldGrade',
  shieldCard = 'shieldCard',
  shieldEnchant1 = 'shieldEnchant1',
  shieldEnchant2 = 'shieldEnchant2',
  shieldEnchant3 = 'shieldEnchant3',

  headUpper = 'headUpper',
  headUpperRefine = 'headUpperRefine',
  headUpperGrade = 'headUpperGrade',
  headUpperCard = 'headUpperCard',
  headUpperEnchant1 = 'headUpperEnchant1',
  headUpperEnchant2 = 'headUpperEnchant2',
  headUpperEnchant3 = 'headUpperEnchant3',
  headMiddle = 'headMiddle',
  headMiddleCard = 'headMiddleCard',
  headMiddleGrade = 'headMiddleGrade',
  headMiddleEnchant1 = 'headMiddleEnchant1',
  headMiddleEnchant2 = 'headMiddleEnchant2',
  headMiddleEnchant3 = 'headMiddleEnchant3',
  headLower = 'headLower',
  headLowerCard = 'headLowerCard',
  headLowerGrade = 'headLowerGrade',
  headLowerEnchant1 = 'headLowerEnchant1',
  headLowerEnchant2 = 'headLowerEnchant2',
  headLowerEnchant3 = 'headLowerEnchant3',
  armor = 'armor',
  armorRefine = 'armorRefine',
  armorGrade = 'armorGrade',
  armorCard = 'armorCard',
  armorEnchant1 = 'armorEnchant1',
  armorEnchant2 = 'armorEnchant2',
  armorEnchant3 = 'armorEnchant3',
  garment = 'garment',
  garmentRefine = 'garmentRefine',
  garmentGrade = 'garmentGrade',
  garmentCard = 'garmentCard',
  garmentEnchant1 = 'garmentEnchant1',
  garmentEnchant2 = 'garmentEnchant2',
  garmentEnchant3 = 'garmentEnchant3',
  boot = 'boot',
  bootRefine = 'bootRefine',
  bootGrade = 'bootGrade',
  bootCard = 'bootCard',
  bootEnchant1 = 'bootEnchant1',
  bootEnchant2 = 'bootEnchant2',
  bootEnchant3 = 'bootEnchant3',
  accLeft = 'accLeft',
  accLeftRefine = 'accLeftRefine',
  accLeftGrade = 'accLeftGrade',
  accLeftCard = 'accLeftCard',
  accLeftEnchant1 = 'accLeftEnchant1',
  accLeftEnchant2 = 'accLeftEnchant2',
  accLeftEnchant3 = 'accLeftEnchant3',
  accRight = 'accRight',
  accRightRefine = 'accRightRefine',
  accRightGrade = 'accRightGrade',
  accRightCard = 'accRightCard',
  accRightEnchant1 = 'accRightEnchant1',
  accRightEnchant2 = 'accRightEnchant2',
  accRightEnchant3 = 'accRightEnchant3',
  pet = 'pet',

  costumeUpper = 'costumeUpper',
  costumeMiddle = 'costumeMiddle',
  costumeLower = 'costumeLower',
  costumeGarment = 'costumeGarment',

  costumeEnchantUpper = 'costumeEnchantUpper',
  costumeEnchantMiddle = 'costumeEnchantMiddle',
  costumeEnchantLower = 'costumeEnchantLower',
  costumeEnchantGarment = 'costumeEnchantGarment',
  costumeEnchantGarment2 = 'costumeEnchantGarment2',
  costumeEnchantGarment4 = 'costumeEnchantGarment4',

  shadowWeapon = 'shadowWeapon',
  shadowWeaponRefine = 'shadowWeaponRefine',
  shadowWeaponEnchant2 = 'shadowWeaponEnchant2',
  shadowWeaponEnchant3 = 'shadowWeaponEnchant3',
  shadowArmor = 'shadowArmor',
  shadowArmorRefine = 'shadowArmorRefine',
  shadowArmorEnchant2 = 'shadowArmorEnchant2',
  shadowArmorEnchant3 = 'shadowArmorEnchant3',
  shadowShield = 'shadowShield',
  shadowShieldRefine = 'shadowShieldRefine',
  shadowShieldEnchant2 = 'shadowShieldEnchant2',
  shadowShieldEnchant3 = 'shadowShieldEnchant3',
  shadowBoot = 'shadowBoot',
  shadowBootRefine = 'shadowBootRefine',
  shadowBootEnchant2 = 'shadowBootEnchant2',
  shadowBootEnchant3 = 'shadowBootEnchant3',
  shadowEarring = 'shadowEarring',
  shadowEarringRefine = 'shadowEarringRefine',
  shadowEarringEnchant2 = 'shadowEarringEnchant2',
  shadowEarringEnchant3 = 'shadowEarringEnchant3',
  shadowPendant = 'shadowPendant',
  shadowPendantRefine = 'shadowPendantRefine',
  shadowPendantEnchant2 = 'shadowPendantEnchant2',
  shadowPendantEnchant3 = 'shadowPendantEnchant3',
}
// ...
export const MainItemWithRelations: Partial<Record<ItemTypeEnum, ItemTypeEnum[]>> = {
  [ItemTypeEnum.weapon]: [
    ItemTypeEnum.weaponCard1,
    ItemTypeEnum.weaponCard2,
    ItemTypeEnum.weaponCard3,
    ItemTypeEnum.weaponCard4,
    ItemTypeEnum.weaponEnchant0,
    ItemTypeEnum.weaponEnchant1,
    ItemTypeEnum.weaponEnchant2,
    ItemTypeEnum.weaponEnchant3,
  ],
  [ItemTypeEnum.ammo]: [],
  [ItemTypeEnum.leftWeapon]: [
    ItemTypeEnum.leftWeaponCard1,
    ItemTypeEnum.leftWeaponCard2,
    ItemTypeEnum.leftWeaponCard3,
    ItemTypeEnum.leftWeaponCard4,
    ItemTypeEnum.leftWeaponEnchant0,
    ItemTypeEnum.leftWeaponEnchant1,
    ItemTypeEnum.leftWeaponEnchant2,
    ItemTypeEnum.leftWeaponEnchant3,
  ],
  [ItemTypeEnum.shield]: [ItemTypeEnum.shieldCard, ItemTypeEnum.shieldEnchant1, ItemTypeEnum.shieldEnchant2, ItemTypeEnum.shieldEnchant3],
  [ItemTypeEnum.headUpper]: [ItemTypeEnum.headUpperCard, ItemTypeEnum.headUpperEnchant1, ItemTypeEnum.headUpperEnchant2, ItemTypeEnum.headUpperEnchant3],
  [ItemTypeEnum.headMiddle]: [ItemTypeEnum.headMiddleCard, ItemTypeEnum.headMiddleEnchant1, ItemTypeEnum.headMiddleEnchant2, ItemTypeEnum.headMiddleEnchant3],
  [ItemTypeEnum.headLower]: [ItemTypeEnum.headLowerCard, ItemTypeEnum.headLowerEnchant1, ItemTypeEnum.headLowerEnchant2, ItemTypeEnum.headLowerEnchant3],
  [ItemTypeEnum.armor]: [ItemTypeEnum.armorCard, ItemTypeEnum.armorEnchant1, ItemTypeEnum.armorEnchant2, ItemTypeEnum.armorEnchant3],
  [ItemTypeEnum.garment]: [ItemTypeEnum.garmentCard, ItemTypeEnum.garmentEnchant1, ItemTypeEnum.garmentEnchant2, ItemTypeEnum.garmentEnchant3],
  [ItemTypeEnum.boot]: [ItemTypeEnum.bootCard, ItemTypeEnum.bootEnchant1, ItemTypeEnum.bootEnchant2, ItemTypeEnum.bootEnchant3],
  [ItemTypeEnum.accRight]: [ItemTypeEnum.accRightCard, ItemTypeEnum.accRightEnchant1, ItemTypeEnum.accRightEnchant2, ItemTypeEnum.accRightEnchant3],
  [ItemTypeEnum.accLeft]: [ItemTypeEnum.accLeftCard, ItemTypeEnum.accLeftEnchant1, ItemTypeEnum.accLeftEnchant2, ItemTypeEnum.accLeftEnchant3],

  [ItemTypeEnum.pet]: [],

  [ItemTypeEnum.costumeUpper]: [],
  [ItemTypeEnum.costumeMiddle]: [],
  [ItemTypeEnum.costumeLower]: [],
  [ItemTypeEnum.costumeGarment]: [],

  [ItemTypeEnum.costumeEnchantUpper]: [],
  [ItemTypeEnum.costumeEnchantMiddle]: [],
  [ItemTypeEnum.costumeEnchantLower]: [],
  [ItemTypeEnum.costumeEnchantGarment]: [],
  [ItemTypeEnum.costumeEnchantGarment2]: [],
  [ItemTypeEnum.costumeEnchantGarment4]: [],

  [ItemTypeEnum.shadowWeapon]: [ItemTypeEnum.shadowWeaponEnchant2, ItemTypeEnum.shadowWeaponEnchant3],
  [ItemTypeEnum.shadowShield]: [ItemTypeEnum.shadowShieldEnchant2, ItemTypeEnum.shadowShieldEnchant3],
  [ItemTypeEnum.shadowArmor]: [ItemTypeEnum.shadowArmorEnchant2, ItemTypeEnum.shadowArmorEnchant3],
  [ItemTypeEnum.shadowBoot]: [ItemTypeEnum.shadowBootEnchant2, ItemTypeEnum.shadowBootEnchant3],
  [ItemTypeEnum.shadowEarring]: [ItemTypeEnum.shadowEarringEnchant2, ItemTypeEnum.shadowEarringEnchant3],
  [ItemTypeEnum.shadowPendant]: [ItemTypeEnum.shadowPendantEnchant2, ItemTypeEnum.shadowPendantEnchant3],
};